+++ frontend/src/component/fileManagement/FileList.tsx
@@ -72,28 +72,6 @@ const FileList: React.FC<Props> = ({
 
 	const [isOpenNewFolder, setIsOpenNewFolder] = useState(false);
 
-	const [anchorPoint, setAnchorPoint] = useState({ x: 0, y: 0 });
-    const [show, setShow] = useState(false);
-
-	const handleContextMenu = useCallback(
-		(e) => {
-		  e.preventDefault();
-		  setAnchorPoint({ x: e.pageX, y: e.pageY });
-		  setShow(true);
-		},
-		[setAnchorPoint, setShow]
-	);
-	const handleClick = useCallback(() => (show ? setShow(false) : null), [show]);
-
-	useEffect(() => {
-		document.addEventListener("click", handleClick);
-		document.addEventListener("contextmenu", handleContextMenu);
-		return () => {
-		  document.removeEventListener("click", handleClick);
-		  document.removeEventListener("contextmenu", handleContextMenu);
-		};
-	});
-
 	return (
 		<Container className={className} ref={container}>
 			<FileHeader>
